* REAL fix

+++ caca/driver_gl.c
@@ -67,7 +67,7 @@ struct driver_private
     cucul_font_t *f;
     float font_width, font_height;
     float incx, incy;
-    unsigned long int const *blocks;
+    uint32_t const *blocks;
     int *txid;
     unsigned char close;
     unsigned char bit;
@@ -498,7 +498,7 @@ static void gl_compute_font(caca_display_t *dp)
     int i, b, w, h, x, y;
 
     /* Count how many glyphs this font has */
-    dp->drv.p->blocks = cucul_get_font_blocks(dp->drv.p->f);
+    dp->drv.p->blocks = (uint32_t)cucul_get_font_blocks(dp->drv.p->f);
 
     for(b = 0, i = 0; dp->drv.p->blocks[i + 1]; i += 2)
         b += (int)(dp->drv.p->blocks[i + 1] - dp->drv.p->blocks[i]);
